Vector: Correct insert() functionality
insert() shifts the elements an index to the right but only once the
index at which we want to insert is encountered. This requires us to
keep track of the element at the current index and the element at the
next index. That way the current index can be updated without us losing
the value that was at the current index.