Vector: Add copy constructor and operator
Copy constructor and operator for a vector. These are different than
many other copy constructor and operators because when are copying
pointers, we can't have the two objects pointing to the same places. If
that happens when we change one object it will the other object.

A class that is responsible for an object accessed through a pointer,
like a vector because vectors are just arrays of pointers, they are
called resource handles. A strong hint that the memberwise copy won't
work is the fact that vector has its own destructor.

For a vector we need both a copy constructor and a copy operator. The
copy constructor will allocate space, assign the default values, and
copy the elements of the other vector into it. This way each vector has
its own copy of the elements.