Deque: Add deque header implementation
Deque is is just a vector with a few extra functions that allow us to
modify the front easier. (secret is, you just insert at 0). Also, I
recognize that good implementations of deque are more than just a few
extra functions on top of a vector but I am fine with what this one is.
For now at least, subject to change.