Vector+TestVector: Allow insertion in correct position when size == 1
When the existing vector had only one element and then we tried to
insert another element at index 0, it would instead just push_back that
element which is not what should happen. To fix this I removed part of
the conditional in the vector insert function.

Added a test to check for this in the future. All tests pass.