Tests Create tests for Set library
Tests all the functions of the Set library as well as their expected
behavior. Things such as trying to insert an element that is already in
the set etc.

Need to create tests to check that the set is sorted but can't do this
until I implement the Iterator library. Sets can't be subscripted and
need to use iterators so at the moment there is no way to iterate
over a set. (data() might work but I haven't given it any thought)