String: Change string to be an extension of vector
std::strings are mutable. Because they are mutable they need to be an
extension of a vector. Typically a string is a "string literal" which
is a char const* which means it is immutable. Because we are wrtiting a
string class we want our strings to be mutable so that we can have
functions like insert, push_back, replace (modifying functions).

I started of DTS::String using char const* but quickly realized that
wasn't going to work due to wanting DTS::String mutable. This PR
switches to using a DTS::Vector with overloaded functions and extra
functions. :)